finished orginazing auth functions
